First successful build

+++ src/libhmg/victory_check.cpp
@@ -3,14 +3,15 @@
 #include <string>
 #include <iostream>
 #include <vector>
+#include <iostream>
 
 using namespace std;
 
 int victory_check(string answer, vector<char>& found)
 {
-    int sum_letters = 0;
-    for (int i = 0; i < found.size(); i++) {
-        for (int j = 0; j < answer.size(); j++) {
+    size_t sum_letters = 0;
+    for (size_t i = 0; i < found.size(); i++) {
+        for (size_t j = 0; j < answer.size(); j++) {
             if (answer.at(j) == found.at(i))
                 sum_letters++;
             if (sum_letters == answer.size())
