+++ src/libhmg/word_check.cpp
@@ -11,7 +11,7 @@ int word_check(string answer, char letter, vector<char>& found)
     for (char i : found)
         if (letter == i)
             return 11;
-    for (int i = 0; i < answer.size(); i++) {
+    for (size_t i = 0; i < answer.size(); i++) {
         if (letter == answer.at(i)) {
             found.push_back(letter);
             return 0;
