+++ src/libhmg/word_fill.cpp
@@ -9,9 +9,9 @@ using namespace std;
 int word_fill(string answer, vector<char>& found)
 {
     bool f = 0;
-    for (int i = 0; i < answer.size(); i++) {
+    for (size_t i = 0; i < answer.size(); i++) {
         f = 0;
-        for (int j = 0; j < found.size(); j++) {
+        for (size_t j = 0; j < found.size(); j++) {
             if (found.at(j) == answer.at(i)) {
                 f = 1;
                 break;
